API query

+++ src/main/java/vttp2022/miniproject/models/ConversionUtils.java
@@ -18,8 +18,9 @@ public class ConversionUtils {
         stock.setSymbol(rs.getString("symbol"));
         stock.setCompany_name(rs.getString("company_name"));
         stock.setShares(rs.getInt("shares"));
-        stock.setShare_price(rs.getFloat("share_price"));
+        stock.setShare_price(rs.getDouble("share_price"));
         stock.setDate_traded(rs.getDate("date_traded"));
+        
         return stock;
     }
 }
